Refactor and optimize code across multiple modules
- Updated chunked upload logic in Azure backend to use `div_ceil` for chunk count calculation.
- Simplified object creation in GCS backend by using struct initialization syntax.
- Removed unnecessary type casting for part numbers in MinIO and S3 backends.
- Improved benchmark tests by eliminating redundant async calls and using more concise assertions.
- Enhanced chunking strategy enum with default variant and removed redundant default implementation.
- Refactored delta encoding logic to use `or_default` for hash table entries.
- Updated merge strategies to use default attributes and removed unnecessary default implementations.
- Improved object type parsing and testing for clarity and consistency.
- Streamlined reflog and refs handling with better error checking.
- Enhanced streaming index and pack writing logic for performance.
- Fixed various test cases to ensure consistency and clarity in commit references.

+++ crates/mediagit-media/src/image.rs
@@ -119,7 +119,7 @@ impl PerceptualHash {
             &img_hash::ImageHash::from_base64(&other.hash_value).ok()?,
         );
 
-        let max_distance = (self.hash_size * self.hash_size) as u32;
+        let max_distance = self.hash_size * self.hash_size;
         Some(1.0 - (hamming_distance as f64 / max_distance as f64))
     }
 }
