+++ crates/mediagit-storage/src/s3.rs
@@ -809,7 +809,7 @@ impl S3Backend {
                     .bucket(&bucket)
                     .key(&key)
                     .upload_id(&upload_id)
-                    .part_number(part_num as i32)
+                    .part_number(part_num)
                     .body(Bytes::from(chunk_data.clone()).into())
                     .send()
                     .await
@@ -853,7 +853,7 @@ impl S3Backend {
             .into_iter()
             .map(|(part_num, etag)| {
                 aws_sdk_s3::types::CompletedPart::builder()
-                    .part_number(part_num as i32)
+                    .part_number(part_num)
                     .e_tag(etag)
                     .build()
             })
